added ability to save network files so that I can test queryDSGRN

+++ shellscripts/networkperturbations_helper_local.sh
@@ -18,6 +18,7 @@ RESULTSDIR=$5
 NETWORKID=$6
 QUERYFILE=$7
 RMDB=$8
+RMNF=$9
 
 DATABASEFILE="$DATABASEDIR/database$NETWORKID.db"
 
@@ -64,8 +65,10 @@ else
 	python pythonmodules/summaryJSON.py $NETWORKFILE "" $RESULTSFILE "$SUMMARYSTR" ""
 fi
 
-# delete intermediate files; it is possible that $STABLEFCLIST does not exist
-rm $NETWORKFILE 
+# delete intermediate files
+if  [  $RMNF == "True" ]; then 
+	rm $NETWORKFILE 
+fi 
 if  [  $RMDB == "True" ]; then 
 	rm $DATABASEFILE
 fi
